backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openpyxl==3.1.5
Pillow==10.4.0
python-multipart==0.0.9
httpx==0.27.2
aiofiles==24.1.0
